Moved module examples
