--- examples/tiles/tiled_example/1/0/0.json
@@ -1,71 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              0,
-              0
-            ],
-            [
-              10000,
-              0
-            ],
-            [
-              10000,
-              10000
-            ],
-            [
-              0,
-              10000
-            ],
-            [
-              0,
-              0
-            ]
-          ]
-        ]
-      },
-      "properties": {
-        "numeric": {
-          "label": 2
-        }
-      }
-    }
-  ],
-  "multiscale": {
-    "axes": [
-      {
-        "name": "x",
-        "type": "space",
-        "unit": "micrometer",
-        "description": "x-axis"
-      },
-      {
-        "name": "y",
-        "type": "space",
-        "unit": "micrometer",
-        "description": "y-axis"
-      }
-    ]
-  },
-  "valueRange": {
-    "label": {
-      "min": 2,
-      "max": 2
-    }
-  },
-  "properties": {
-    "string": {
-      "plate": "label",
-      "image": "x00_y01_p01_c1.ome.tif"
-    },
-    "numeric": {
-      "channel": 1.0
-    }
-  }
-}
