--- examples/tiles/tiled_example/2/2/3.json
@@ -1,71 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              10000,
-              15000
-            ],
-            [
-              15000,
-              15000
-            ],
-            [
-              15000,
-              20000
-            ],
-            [
-              10000,
-              20000
-            ],
-            [
-              10000,
-              15000
-            ]
-          ]
-        ]
-      },
-      "properties": {
-        "numeric": {
-          "label": 17
-        }
-      }
-    }
-  ],
-  "multiscale": {
-    "axes": [
-      {
-        "name": "x",
-        "type": "space",
-        "unit": "micrometer",
-        "description": "x-axis"
-      },
-      {
-        "name": "y",
-        "type": "space",
-        "unit": "micrometer",
-        "description": "y-axis"
-      }
-    ]
-  },
-  "valueRange": {
-    "label": {
-      "min": 17,
-      "max": 17
-    }
-  },
-  "properties": {
-    "string": {
-      "plate": "label",
-      "image": "x00_y01_p01_c1.ome.tif"
-    },
-    "numeric": {
-      "channel": 1.0
-    }
-  }
-}
