+++ mypy.ini
@@ -1,3 +1,4 @@
 [mypy]
 disallow_any_generics = False
 exclude = src/microjson/utils.py
+ignore_missing_imports = True
