+++ src/microjson/examples/tiles/tiled_example/2/1/1.json
@@ -1,0 +1,71 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              5000,
+              5000
+            ],
+            [
+              10000,
+              5000
+            ],
+            [
+              10000,
+              10000
+            ],
+            [
+              5000,
+              10000
+            ],
+            [
+              5000,
+              5000
+            ]
+          ]
+        ]
+      },
+      "properties": {
+        "numeric": {
+          "label": 11
+        }
+      }
+    }
+  ],
+  "multiscale": {
+    "axes": [
+      {
+        "name": "x",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "x-axis"
+      },
+      {
+        "name": "y",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "y-axis"
+      }
+    ]
+  },
+  "valueRange": {
+    "label": {
+      "min": 11,
+      "max": 11
+    }
+  },
+  "properties": {
+    "string": {
+      "plate": "label",
+      "image": "x00_y01_p01_c1.ome.tif"
+    },
+    "numeric": {
+      "channel": 1.0
+    }
+  }
+}
