+++ src/microjson/examples/tiles/tiled_example/2/3/0.json
@@ -1,0 +1,71 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              15000,
+              0
+            ],
+            [
+              20000,
+              0
+            ],
+            [
+              20000,
+              5000
+            ],
+            [
+              15000,
+              5000
+            ],
+            [
+              15000,
+              0
+            ]
+          ]
+        ]
+      },
+      "properties": {
+        "numeric": {
+          "label": 18
+        }
+      }
+    }
+  ],
+  "multiscale": {
+    "axes": [
+      {
+        "name": "x",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "x-axis"
+      },
+      {
+        "name": "y",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "y-axis"
+      }
+    ]
+  },
+  "valueRange": {
+    "label": {
+      "min": 18,
+      "max": 18
+    }
+  },
+  "properties": {
+    "string": {
+      "plate": "label",
+      "image": "x00_y01_p01_c1.ome.tif"
+    },
+    "numeric": {
+      "channel": 1.0
+    }
+  }
+}
