+++ src/microjson/examples/tiles/tiled_example/2/3/2.json
@@ -1,0 +1,71 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              15000,
+              10000
+            ],
+            [
+              20000,
+              10000
+            ],
+            [
+              20000,
+              15000
+            ],
+            [
+              15000,
+              15000
+            ],
+            [
+              15000,
+              10000
+            ]
+          ]
+        ]
+      },
+      "properties": {
+        "numeric": {
+          "label": 20
+        }
+      }
+    }
+  ],
+  "multiscale": {
+    "axes": [
+      {
+        "name": "x",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "x-axis"
+      },
+      {
+        "name": "y",
+        "type": "space",
+        "unit": "micrometer",
+        "description": "y-axis"
+      }
+    ]
+  },
+  "valueRange": {
+    "label": {
+      "min": 20,
+      "max": 20
+    }
+  },
+  "properties": {
+    "string": {
+      "plate": "label",
+      "image": "x00_y01_p01_c1.ome.tif"
+    },
+    "numeric": {
+      "channel": 1.0
+    }
+  }
+}
